refactor(types): centralize BaseEntity/FirestoreId in @models/firestore and wire aliases\n\n- add src/types/firestore.ts and @models alias (vite/vitest/tsconfig)\n- update types and useFirestore to import shared types
