+++ src/types/customer.ts
@@ -1,4 +1,4 @@
-import type { BaseEntity } from "@hooks/useFirestore";
+import type { BaseEntity } from "@models/firestore";
 
 export type Customer = BaseEntity & {
   userId: string;
