+++ src/types/item.ts
@@ -1,4 +1,4 @@
-import type { BaseEntity } from "@hooks/useFirestore";
+import type { BaseEntity } from "@models/firestore";
 
 export type Item = BaseEntity & {
   userId: string;
