+++ vite.config.ts
@@ -14,6 +14,7 @@ export default defineConfig({
       "@utils": fileURLToPath(new URL("./src/utils", import.meta.url)),
       "@hooks": fileURLToPath(new URL("./src/hooks", import.meta.url)),
       "@auth": fileURLToPath(new URL("./src/auth", import.meta.url)),
+      "@models": fileURLToPath(new URL("./src/types", import.meta.url)),
     },
   },
   // Removed COOP/COEP headers to allow loading cross-origin resources like CDN scripts during development
